Refactored builder
