+++ .env
@@ -110,7 +110,6 @@ ESPRESSO_RESERVE_BUILDER_SERVER_PORT=31005
 ESPRESSO_BUILDER_TX_CHANNEL_CAPACITY=4096
 ESPRESSO_BUILDER_EVENT_CHANNEL_CAPACITY=128
 ESPRESSO_BUILDER_INIT_NODE_COUNT=$ESPRESSO_ORCHESTRATOR_NUM_NODES
-ESPRESSO_BUILDER_BOOTSTRAPPED_VIEW=0
 ESPRESSO_BUILDER_WEBSERVER_RESPONSE_TIMEOUT_DURATION=1500ms
 ESPRESSO_BUILDER_BUFFER_VIEW_NUM_COUNT=50
 ESPRESSO_BUILDER_GENESIS_FILE=$ESPRESSO_SEQUENCER_GENESIS_FILE
@@ -149,5 +148,5 @@ INTEGRATION_TEST_PROTO=http
 # `03`, marketplace upgrade will be tested.
 INTEGRATION_TEST_SEQUENCER_VERSION=02
 
-# max database connections 
+# max database connections
 ESPRESSO_SEQUENCER_DATABASE_MAX_CONNECTIONS=25
