+++ builder/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "builder"
+name = "permissionless-builder"
 description = "A standalone builder service"
 version = { workspace = true }
 authors = { workspace = true }
