+++ docker-compose.yaml
@@ -542,7 +542,6 @@ services:
       - ESPRESSO_BUILDER_SERVER_PORT
       - ESPRESSO_BUILDER_TX_CHANNEL_CAPACITY
       - ESPRESSO_BUILDER_EVENT_CHANNEL_CAPACITY
-      - ESPRESSO_BUILDER_BOOTSTRAPPED_VIEW
       - ESPRESSO_BUILDER_WEBSERVER_RESPONSE_TIMEOUT_DURATION
       - ESPRESSO_BUILDER_BUFFER_VIEW_NUM_COUNT
       - ESPRESSO_BUILDER_INIT_NODE_COUNT
@@ -573,7 +572,6 @@ services:
       - ESPRESSO_BUILDER_L1_PROVIDER
       - ESPRESSO_BUILDER_TX_CHANNEL_CAPACITY
       - ESPRESSO_BUILDER_EVENT_CHANNEL_CAPACITY
-      - ESPRESSO_BUILDER_BOOTSTRAPPED_VIEW
       - ESPRESSO_BUILDER_WEBSERVER_RESPONSE_TIMEOUT_DURATION
       - ESPRESSO_BUILDER_BUFFER_VIEW_NUM_COUNT
       - ESPRESSO_BUILDER_INIT_NODE_COUNT
@@ -600,7 +598,6 @@ services:
       - ESPRESSO_BUILDER_SERVER_PORT
       - ESPRESSO_BUILDER_TX_CHANNEL_CAPACITY
       - ESPRESSO_BUILDER_EVENT_CHANNEL_CAPACITY
-      - ESPRESSO_BUILDER_BOOTSTRAPPED_VIEW
       - ESPRESSO_BUILDER_WEBSERVER_RESPONSE_TIMEOUT_DURATION
       - ESPRESSO_BUILDER_BUFFER_VIEW_NUM_COUNT
       - ESPRESSO_BUILDER_INIT_NODE_COUNT
@@ -733,8 +730,7 @@ services:
 
   dev-rollup:
     image: ghcr.io/espressosystems/espresso-sequencer/dev-rollup:main
-    command:
-      dev-rollup register --ns 1; dev-rollup register --ns 2; dev-rollup register --ns 3
+    command: dev-rollup register --ns 1; dev-rollup register --ns 2; dev-rollup register --ns 3
     environment:
       - ESPRESSO_MARKETPLACE_SOLVER_API_URL=http://marketplace-solver:$ESPRESSO_MARKETPLACE_SOLVER_API_PORT
     depends_on:
